refactor: separate manager initialization from IPC dispatching
Extract manager-init.js to handle inter-manager dependency wiring
(flowManager.start, sessionManager.start, usageManager.init) and
target map construction. ipc-handlers.js now receives a pre-built
targets object via function parameter, reducing its imports from 7
managers down to zero and making the initialization order explicit
in the main entry point.
