+++ main.js
@@ -1,12 +1,20 @@
 const { app } = require('electron');
 const window = require('./main/window');
+const { initManagers } = require('./main/manager-init');
 const ipcHandlers = require('./main/ipc-handlers');
 
 app.setName('Pickagent');
 
+let managerCleanup = null;
+
 app.whenReady().then(() => {
   const win = window.create();
-  ipcHandlers.register(() => window.get());
+  const getWindow = () => window.get();
+
+  const { targets, cleanup, ptyManager, sessionManager } = initManagers(getWindow);
+  managerCleanup = cleanup;
+
+  ipcHandlers.register(getWindow, { targets, ptyManager, sessionManager });
 
   app.on('activate', () => {
     if (!window.get()) window.create();
@@ -14,6 +22,6 @@ app.whenReady().then(() => {
 });
 
 app.on('window-all-closed', () => {
-  ipcHandlers.cleanup();
+  if (managerCleanup) managerCleanup();
   if (process.platform !== 'darwin') app.quit();
 });
